% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfun.R
\name{bootstrapfun}
\alias{bootstrapfun}
\title{Bootstrap for the bivariate copula models}
\usage{
bootstrapfun(Q, family, tau, n, df, max_iter, eps, HMM)
}
\arguments{
\item{Q}{Weights vector (1 x reg or component);}

\item{family}{'gaussian' , 't' , 'clayton' , 'frank' , 'gumbel'}

\item{tau}{Kendall's rank correlation}

\item{n}{number of simulated vectors}

\item{eps}{precision (e.g 0.00001);}

\item{HMM}{1 (if HMM) , 0 (if mixture);}

\item{DoF}{vector of degree of freedom (d x 1), only for the Student copula.}

\item{n_sample}{number of bootstrap (e.g 1000);}
}
\description{
Bootstrap for the bivariate copula models
}
\author{
Mamadou Yamar Thioub and Bruno Remillard, April 12, 2018
}
\keyword{internal}
