% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulaFamiliesPDF.R
\name{copulaFamiliesPDF}
\alias{copulaFamiliesPDF}
\title{COPULAPDF Probability density function for a copula.

COPULAPDF  probability density function for a copula with linear correlation parameters RHO and}
\usage{
copulaFamiliesPDF(family, u, ...)
}
\arguments{
\item{family}{copula familly= "gaussian" , "t" , "clayton" , "frank" , "gumbel"}

\item{u}{is an N-by-P matrix of values in [0,1], representing N points in the P-dimensional unit hypercube}

\item{...}{additionnal parameter like RHO  a P-by-P correlation matrix.}
}
\value{
Y = COPULAPDF('Gaussian',U,RHO) returns the probability density of the
  Gaussian copula with linear correlation parameters RHO, evaluated at the
  points in U. U is an N-by-P matrix of values in [0,1], representing N
  points in the P-dimensional unit hypercube.  RHO is a P-by-P correlation
  matrix.  If U is an N-by-2 matrix, RHO may also be a scalar correlation
  coefficient.

  Y = COPULAPDF('t',U,RHO,NU) returns the probability density of the t
  copula with linear correlation parameters RHO and degrees of freedom
  parameter NU, evaluated at the points in U.  U is an N-by-P matrix of
  values in [0,1]. RHO is a P-by-P correlation matrix.  If U is an N-by-2
  matrix, RHO may also be a scalar correlation coefficient.

  Y = COPULAPDF(FAMILY,U,ALPHA) returns the probability density of the
  bivariate Archimedean copula determined by FAMILY, with scalar parameter
  ALPHA, evaluated at the points in U.  FAMILY is 'Clayton', 'Frank', or
  'Gumbel'.  U is an N-by-2 matrix of values in [0,1].
}
\description{
COPULAPDF Probability density function for a copula.

COPULAPDF  probability density function for a copula with linear correlation parameters RHO and
}
\examples{
 u = seq(0,1,0.1);
    U1=matrix(rep(u,length(u)),nrow=length(u),byrow = TRUE); U2=t(U1)
   F = copulaFamiliesPDF('clayton',cbind(c(U1), c(U2)),1)

}
\keyword{internal}
