% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstKendallTau.R
\name{EstKendallTau}
\alias{EstKendallTau}
\title{Sample Kendall's tau Estimation}
\usage{
EstKendallTau(X)
}
\arguments{
\item{X}{(n x 2) matrix}
}
\value{
\item{KendallTau}{estimated sample Kendall's tau of the data}
}
\description{
This function estimates the sample Kendall's tau of a bivariate data matrix
}
