% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallTau.R
\name{KendallTau}
\alias{KendallTau}
\title{Kendall's tau of a copula}
\usage{
KendallTau(family, alpha)
}
\arguments{
\item{family}{"gaussian" , "t" , "clayton" , "frank" , "gumbel"}

\item{alpha}{unconstrainted parameters of the copula family}
}
\value{
\item{tau}{estimated Kendall's tau}
}
\description{
This function computes the Kendall's tau of a copula family with a unconstrainted parameter alpha.
}
