% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamCop.R
\name{ParamCop}
\alias{ParamCop}
\title{Theta estimation}
\usage{
ParamCop(family, alpha)
}
\arguments{
\item{family}{"gaussian" , "t" , "clayton" , "frank" , "gumbel"}

\item{alpha}{unconstrainted parameters of the copula family}
}
\value{
\item{theta}{matlab parameters}
}
\description{
Parameters of a copula according to CRAN copula package (except for Frank copula, where theta = log(theta_R_Package)), corresponding to the unconstrainted parameters alpha.
}
