% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTau.R
\name{ParamTau}
\alias{ParamTau}
\title{Alpha estimation}
\usage{
ParamTau(family, tau)
}
\arguments{
\item{family}{'gaussian' , 't' , 'clayton' , 'frank' , 'gumbel'}

\item{tau}{Kendall's tau of the copula family}
}
\value{
\item{alpha}{estimated unconstrainted parameter}
}
\description{
Unconstrainted parameter for a given Kendall's tau.
}
