% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosenblattClayton.R
\name{RosenblattClayton}
\alias{RosenblattClayton}
\title{Rosenblatt transform for Clayton copula}
\usage{
RosenblattClayton(u, theta)
}
\arguments{
\item{u}{(n x d) matrix of pseudos-observations (normalized ranks)}

\item{theta}{parameter of the Clayton copula}
}
\value{
\item{R}{Rosenblatt transform}
}
\description{
This function computes the Rosenblatt transform fot the Clayton copula
}
