% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosenblattFrank.R
\name{RosenblattFrank}
\alias{RosenblattFrank}
\title{Rosenblatt transform for Frank copula}
\usage{
RosenblattFrank(U, theta)
}
\arguments{
\item{U}{(n x d) matrix of pseudos-observations (normalized ranks)}

\item{theta}{parameter of the Frank copula}
}
\value{
\item{R}{Rosenblatt transform}
}
\description{
Computation of the Rosenblatt transform for Frank's copula
}
