% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosenblattGaussian.R
\name{RosenblattGaussian}
\alias{RosenblattGaussian}
\title{Rosenblatt transform for Gaussian copula}
\usage{
RosenblattGaussian(u, rho)
}
\arguments{
\item{u}{(n x d) matrix of pseudos-observations (normalized ranks)}

\item{rho}{(d x d) correlation matrix, or the correlation coefficient (if, d = 2)}
}
\value{
\item{R}{Rosenblatt transform}
}
\description{
Computation of the Rosenblatt transform for the Gaussian copula
}
