% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosenblattStudent.R
\name{RosenblattStudent}
\alias{RosenblattStudent}
\title{Rosenblatt transform for Student copula}
\usage{
RosenblattStudent(u, rho, nu)
}
\arguments{
\item{u}{(n x d) matrix of pseudos-observations (normalized ranks)}

\item{rho}{(d x d) correlation matrix}

\item{nu}{degrees of freedom}
}
\value{
\item{R}{Rosenblatt transform}
}
\description{
This function computes the Rosenblatt transform fot the Student copula
}
