% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnB.R
\name{SnB}
\alias{SnB}
\title{Cramer-von Mises statistic SnB for GOF based on the Rosenblatt transform}
\usage{
SnB(E)
}
\arguments{
\item{E}{(n x d) matrix of pseudos-observations (normalized ranks)}
}
\value{
\item{Sn}{Cramer-von Mises statistic}
}
\description{
Computation of the Cramer-von Mises statistic SnB for GOF based on the Rosenblatt transform
}
