% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tau2Rho.R
\name{Tau2Rho}
\alias{Tau2Rho}
\title{Spearman's rho}
\usage{
Tau2Rho(family, theta)
}
\arguments{
\item{family}{'gaussian' , 't' , 'clayton' , 'frank' , 'gumbel'}

\item{theta}{parameter of the copula according to CRAN copula package (except for Frank copula, where theta = log(theta_R_Package))}
}
\value{
\item{rho}{estimated Spearman's rho}
}
\description{
This function estimates the Spearman's rho corresponding to a constrainted (matlab) parameter theta for a copula family.
}
