% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilog.R
\name{dilog}
\alias{dilog}
\title{Dilogarithm function}
\usage{
dilog(x)
}
\arguments{
\item{x}{a real number}
}
\value{
\item{out}{dilogarithm}
}
\description{
This function computes the dilogarithm of a number.
}
