% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimMarkovChain.R
\name{SimMarkovChain}
\alias{SimMarkovChain}
\title{Markov chain simulation}
\usage{
SimMarkovChain(Q, n, eta0)
}
\arguments{
\item{Q}{Transition probality matrix (d x d)}

\item{n}{number of simulated vectors}

\item{eta0}{variable eta}
}
\description{
This function generates a Markov chain X(1), ..., X(n) with transition matrix Q,
starting from a state eta0 or the uniform distribution on {1,..., r}
}
